Extract settings surface contract
