+++ docs/DECISIONS.md
@@ -145,6 +145,7 @@
 - Transcript message controls live in `QuillCodeTranscriptMessageView.swift`. User and assistant message bubbles, retry/use-as-draft actions, feedback actions, and the shared transcript copy button should evolve there; `WorkspaceSwiftUIView.swift` should only decide where timeline items are placed. Tool-card and artifact-preview rendering remains the next transcript family to extract.
 - Tool-card rendering lives in `QuillCodeToolCardView.swift`. Tool cards, execution chips and rails, artifact chips/previews, and raw detail blocks should evolve there; `WorkspaceSwiftUIView.swift` should only decide where timeline items are placed and how copy actions are wired.
 - Settings and runtime issue rendering live in `QuillCodeSettingsView.swift`. Authentication settings, developer override controls, Computer Use permission onboarding, diagnostics, and `QuillCodeSettingsDraft` should evolve there; `WorkspaceSwiftUIView.swift` should only present the sheet and apply the resulting settings update.
+- Settings surface contracts live in `QuillCodeSettingsSurface.swift`. `WorkspaceSettingsSurface`, `WorkspaceSettingsUpdate`, Computer Use requirement rows, TrustedRouter sign-in copy, Computer Use permission labels, and compatibility decoding stay beside the settings contract; `WorkspaceSurface.swift` should only carry the aggregate `settings` payload and pass current runtime state into it.
 - Terminal and browser utility panes live in `QuillCodeTerminalBrowserPaneView.swift`. Terminal history rows, browser navigation, page snapshots, comments, and pane-local draft state should evolve there; `WorkspaceSwiftUIView.swift` should only decide when the panes are visible and route the resulting actions.
 - Secondary utility panes live in `QuillCodeSecondaryPanesView.swift`. Extensions, Memories, Automations, their count/status pills, empty states, and compact cards should evolve there; `WorkspaceSwiftUIView.swift` should only decide pane placement and route commands.
 - Workspace modal surfaces live in `QuillCodeWorkspaceDialogs.swift`. Command palette, keyboard shortcuts, chat search, rename sheets, and worktree create/remove sheets should evolve there; `WorkspaceSwiftUIView.swift` should only decide when a sheet is presented and route completed actions.
